+++ test/test_css_parser_misc.rb
@@ -195,20 +195,20 @@ class CssParserTests < Minitest::Test
     # rulesets = []
     #
     # parser['div'].each do |declaration|
-    #   rulesets << RuleSet.new('div', declaration)
+    #   rulesets << RuleSet.new(selectors: 'div', block: declaration)
     # end
     #
     # merged = CssParser.merge(rulesets)
     #
     # result: # merged.to_s => "{ background-color: black !important; }"
 
-    new_rule = RuleSet.new('div', "{ background-color: black !important; }")
+    new_rule = RuleSet.new(selectors: 'div', block: "{ background-color: black !important; }")
 
     assert_equal 'div { background-color: black !important; }', new_rule.to_s
   end
 
   def test_content_with_data
-    rule = RuleSet.new('div', '{content: url(data:image/png;base64,LOTSOFSTUFF)}')
+    rule = RuleSet.new(selectors: 'div', block: '{content: url(data:image/png;base64,LOTSOFSTUFF)}')
     assert_includes rule.to_s, "image/png;base64,LOTSOFSTUFF"
   end
 
@@ -225,5 +225,49 @@ class CssParserTests < Minitest::Test
       assert_equal 'body', sel
       assert_equal 'color: black;', desc
     end
+  end
+
+  def test_catching_argument_exceptions_for_add_rule
+    cp_with_exceptions = Parser.new(rule_set_exceptions: true)
+    assert_raises ArgumentError, 'background-color value is empty' do
+      cp_with_exceptions.add_rule!(selectors: 'body', block: 'background-color: !important')
+    end
+
+    cp_without_exceptions = Parser.new(rule_set_exceptions: false)
+    cp_without_exceptions.add_rule!(selectors: 'body', block: 'background-color: !important')
+  end
+
+  def test_catching_argument_exceptions_for_add_rule_positional
+    cp_with_exceptions = Parser.new(rule_set_exceptions: true)
+
+    assert_raises ArgumentError, 'background-color value is empty' do
+      _, err = capture_io do
+        cp_with_exceptions.add_rule!('body', 'background-color: !important')
+      end
+      assert_includes err, "DEPRECATION"
+    end
+
+    cp_without_exceptions = Parser.new(rule_set_exceptions: false)
+    _, err = capture_io do
+      cp_without_exceptions.add_rule!('body', 'background-color: !important')
+    end
+    assert_includes err, "DEPRECATION"
+  end
+
+  def test_catching_argument_exceptions_for_add_rule_with_offsets
+    cp_with_exceptions = Parser.new(capture_offsets: true, rule_set_exceptions: true)
+
+    assert_raises ArgumentError, 'background-color value is empty' do
+      _, err = capture_io do
+        cp_with_exceptions.add_rule_with_offsets!('body', 'background-color: !important', 'inline', 1)
+      end
+      assert_includes err, "DEPRECATION"
+    end
+
+    cp_without_exceptions = Parser.new(capture_offsets: true, rule_set_exceptions: false)
+    _, err = capture_io do
+      cp_without_exceptions.add_rule_with_offsets!('body', 'background-color: !important', 'inline', 1)
+    end
+    assert_includes err, "DEPRECATION"
   end
 end