+++ test/test_rule_set.rb
@@ -12,7 +12,7 @@ class RuleSetTests < Minitest::Test
   end
 
   def test_setting_property_values
-    rs = RuleSet.new(nil, nil)
+    rs = RuleSet.new
 
     rs['background-color'] = 'red'
     assert_equal('red;', rs['background-color'])
@@ -22,12 +22,12 @@ class RuleSetTests < Minitest::Test
   end
 
   def test_getting_property_values
-    rs = RuleSet.new('#content p, a', 'color: #fff;')
+    rs = RuleSet.new(selectors: '#content p, a', block: 'color: #fff;')
     assert_equal('#fff;', rs['color'])
   end
 
   def test_getting_property_value_ignoring_case
-    rs = RuleSet.new('#content p, a', 'color: #fff;')
+    rs = RuleSet.new(selectors: '#content p, a', block: 'color: #fff;')
     assert_equal('#fff;', rs['  ColoR '])
   end
 
@@ -38,7 +38,7 @@ class RuleSetTests < Minitest::Test
     ]
 
     actual = []
-    rs = RuleSet.new('#content p, a', 'color: #fff;')
+    rs = RuleSet.new(selectors: '#content p, a', block: 'color: #fff;')
     rs.each_selector do |sel, decs, spec|
       actual << {selector: sel, declarations: decs, specificity: spec}
     end
@@ -54,7 +54,7 @@ class RuleSetTests < Minitest::Test
     ]
 
     actual = Set.new
-    rs = RuleSet.new(nil, 'color: #fff; Background: white none no-repeat !important; margin: 1px -0.25em;')
+    rs = RuleSet.new(block: 'color: #fff; Background: white none no-repeat !important; margin: 1px -0.25em;')
     rs.each_declaration do |prop, val, imp|
       actual << {property: prop, value: val, is_important: imp}
     end
@@ -64,7 +64,7 @@ class RuleSetTests < Minitest::Test
 
   def test_each_declaration_respects_order
     css_fragment = "margin: 0; padding: 20px; margin-bottom: 28px;"
-    rs           = RuleSet.new(nil, css_fragment)
+    rs           = RuleSet.new(block: css_fragment)
     expected     = %w[margin padding margin-bottom]
     actual       = []
     rs.each_declaration { |prop, _val, _imp| actual << prop }
@@ -72,38 +72,38 @@ class RuleSetTests < Minitest::Test
   end
 
   def test_each_declaration_containing_semicolons
-    rs = RuleSet.new(nil, "background-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABwAAAAiCAMAAAB7);" \
-                          "background-repeat: no-repeat")
+    rs = RuleSet.new(block: "background-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABwAAAAiCAMAAAB7);" \
+                            "background-repeat: no-repeat")
     assert_equal('url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABwAAAAiCAMAAAB7);', rs['background-image'])
     assert_equal('no-repeat;', rs['background-repeat'])
   end
 
   def test_selector_sanitization
     selectors = "h1, h2,\nh3 "
-    rs = RuleSet.new(selectors, "color: #fff;")
+    rs = RuleSet.new(selectors: selectors, block: "color: #fff;")
     assert rs.selectors.member?("h3")
   end
 
   def test_multiple_selectors_to_s
     selectors = "#content p, a"
-    rs = RuleSet.new(selectors, "color: #fff;")
+    rs = RuleSet.new(selectors: selectors, block: "color: #fff;")
     assert_match(/^\s*#content p,\s*a\s*\{/, rs.to_s)
   end
 
   def test_declarations_to_s
     declarations = 'color: #fff; font-weight: bold;'
-    rs = RuleSet.new('#content p, a', declarations)
+    rs = RuleSet.new(selectors: '#content p, a', block: declarations)
     assert_equal(declarations.split.sort, rs.declarations_to_s.split.sort)
   end
 
   def test_important_declarations_to_s
     declarations = 'color: #fff; font-weight: bold !important;'
-    rs = RuleSet.new('#content p, a', declarations)
+    rs = RuleSet.new(selectors: '#content p, a', block: declarations)
     assert_equal(declarations.split.sort, rs.declarations_to_s.split.sort)
   end
 
   def test_overriding_specificity
-    rs = RuleSet.new('#content p, a', 'color: white', 1000)
+    rs = RuleSet.new(selectors: '#content p, a', block: 'color: white', specificity: 1000)
     rs.each_selector do |_sel, _decs, spec|
       assert_equal 1000, spec
     end
@@ -112,7 +112,7 @@ class RuleSetTests < Minitest::Test
   def test_not_raised_issue68
     ok = true
     begin
-      RuleSet.new('td', 'border-top: 5px solid; border-color: #fffff0;')
+      RuleSet.new(selectors: 'td', block: 'border-top: 5px solid; border-color: #fffff0;')
     rescue
       ok = false
     end